requirements.txt
opencv_python==4.8.1.78
pygame-ce==2.3.2
pipreqs==0.4.13